* Added prelimary support for darwin (MacOSX) * Added -std=89 and other flags * Some general makefile tidying * fixed void* -> int cast to use intptr_t which gave warnings on amd64 * to support the above the configure script checks for intptr_t * * qdata.h:    #include <stdint.h> * models.c:   changed (int) cast to (intptr_t) * images.c:   same

+++ src/tools/qdata/images.c
@@ -583,7 +583,7 @@ void Cmd_Mip (void)
 	}
 
 	/* dword align the size */
-	while ((int)lump_p&3)
+	while ((intptr_t)lump_p&3)
 		*lump_p++ = 0;
 
 	/* write it out */
